Update hpc_tiu.md
Students cannot access the High Performance Computing environment, as it is only available to Tilburg University staff members. Since students do not have staff accounts, they cannot view the request form on the intranet throught this page. This may cause confusion, but please be aware that the HPC environment is not intended for student use.